ScrapeProducts/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
attrs==22.1.0; python_full_version >= "3.7.1" and python_version >= "3.7" \
automat==22.10.0; python_full_version >= "3.7.1" and python_version >= "3.7" \
azure-core==1.26.1; python_version >= "3.7" \
azure-identity==1.12.0; python_version >= "3.7" \
azure-storage-blob==12.14.1; python_version >= "3.7" \
certifi==2022.9.24; python_version >= "3.7" and python_version < "4" \
cffi==1.15.1; python_version >= "3.7" \
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0" \
constantly==15.1.0; python_full_version >= "3.7.1" and python_version >= "3.7" \
cryptography==38.0.3; python_version >= "3.7" \
cssselect==1.2.0; python_version >= "3.7" \
filelock==3.8.0; python_version >= "3.7" \
hyperlink==21.0.0; python_full_version >= "3.7.1" and python_version >= "3.7" \
idna==3.4; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.7.1" \
incremental==22.10.0; python_full_version >= "3.7.1" and python_version >= "3.7" \
isodate==0.6.1; python_version >= "3.7" \
itemadapter==0.7.0; python_version >= "3.7" \
itemloaders==1.0.6; python_version >= "3.7" \
jmespath==1.0.1; python_version >= "3.7" \
lxml==4.9.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7" \
msal-extensions==1.0.0; python_version >= "3.7" \
msal==1.20.0; python_version >= "3.7" \
msrest==0.7.1; python_version >= "3.7" \
oauthlib==3.2.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7" \
packaging==21.3; python_version >= "3.7" \
parsel==1.7.0; python_version >= "3.7" \
portalocker==2.6.0 \
protego==0.2.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7" \
pyasn1-modules==0.2.8; python_version >= "3.7" \
pyasn1==0.4.8; python_version >= "3.7" \
pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7" \
pydispatcher==2.0.6; platform_python_implementation == "CPython" and python_version >= "3.7" \
pyjwt==2.6.0; python_version >= "3.7" \
pyopenssl==22.1.0; python_version >= "3.7" \
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.7" \
pypydispatcher==2.1.2; platform_python_implementation == "PyPy" and python_version >= "3.7" \
pywin32==305; python_version >= "3.7" and platform_system == "Windows" \
queuelib==1.6.2; python_version >= "3.7" \
requests-file==1.5.1; python_version >= "3.7" \
requests-oauthlib==1.3.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7" \
requests==2.28.1; python_version >= "3.7" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7") \
scrapy==2.7.1; python_version >= "3.7" \
service-identity==21.1.0; python_version >= "3.7" \
six==1.16.0; python_full_version >= "3.7.1" and python_version >= "3.7" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7") \
tldextract==3.4.0; python_version >= "3.7" \
twisted-iocpsupport==1.0.2; python_full_version >= "3.7.1" and python_version >= "3.7" and platform_system == "Windows" \
twisted==22.10.0; python_full_version >= "3.7.1" and python_version >= "3.7" \
typing-extensions==4.4.0; python_full_version >= "3.7.1" and python_version >= "3.7" \
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7" \
w3lib==2.0.1; python_version >= "3.7" \
zope.interface==5.5.2; python_full_version >= "3.7.1" and python_version >= "3.7" \